Update Sidebar hooks to work with MW 1.39+
The SkinBuildSidebar no longer accepts raw HTML, and will fail if
it's not passed an array.  There's a different hook added in 1.35,
SkinAfterPortlet, which places HTML content after a portlet.
But we still need the first hook to create a portlet, if it's given.

Currently what will happen if you use an existing portlet name
is that Mediawiki will just overwrite the empty array, which means
you can also append to an existing portlet.